@@ -4,13 +4,14 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
+    "prestart": "npm run config",
     "start": "if-env NODE_ENV=production && npm run -s serve || npm run -s dev",
-    "build": "preact build --no-prerender",
+    "build": "preact build --no-prerender --template src/template.html",
     "config": "cross-env NODE_ENV=dev NODE_ENV_MODE=config tasks/setupConfig.js",
     "deploy": "cross-env NODE_ENV=dist tasks/deploy.js",
-    "serve": "preact build --no-prerender && preact serve",
-    "dev": "preact watch -p 3000 --no-prerender",
-    "lint": "eslint src --quiet",
+    "serve": "preact build --no-prerender --template src/template.html && preact serve",
+    "dev": "preact watch -p 3000 --no-prerender --template src/template.html",
+    "lint": "eslint src --quiet --cache",
     "test": "cross-env NODE_ENV=test jest ./tests",
     "pretty": "prettier -c --write $(find src -type f -name '*.js')"
   },
@@ -63,42 +64,43 @@
     "babel-plugin-transform-react-jsx": "^6.24.1",
     "babel-plugin-transform-runtime": "^6.23.0",
     "chalk": "^2.4.1",
+    "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^5.2.0",
     "dedent": "^0.7.0",
     "del": "^3.0.0",
-    "eslint": "^5.6.1",
+    "eslint": "^5.8.0",
     "eslint-config-standard-preact": "^1.1.6",
     "eslint-config-synacor": "^3.0.3",
     "execa": "^1.0.0",
-    "husky": "^1.1.0",
+    "husky": "^1.1.3",
     "identity-obj-proxy": "^3.0.0",
     "if-env": "^1.0.0",
     "jest": "^23.6.0",
-    "jest-localstorage-mock": "^2.2.0",
+    "jest-localstorage-mock": "^2.3.0",
     "jest-serializer-html-string": "^1.0.1",
-    "lint-staged": "^7.3.0",
+    "lint-staged": "^8.0.4",
     "listr": "^0.14.1",
-    "listr-update-renderer": "^0.4.0",
+    "listr-update-renderer": "^0.5.0",
     "minimist": "^1.2.0",
-    "preact-cli": "^2.1.0",
+    "preact-cli": "^3.0.0-next.14",
     "preact-render-spy": "^1.2.1",
-    "preact-render-to-string": "^3.7.0",
-    "prettier": "^1.14.3"
+    "preact-render-to-string": "^4.1.0",
+    "prettier": "^1.15.3"
   },
   "dependencies": {
-    "c-3po": "^0.8.1",
-    "lodash": "^4.17.11",
-    "file-saver": "^1.3.8",
-    "moment": "^2.22.2",
+    "file-saver": "^2.0.0",
     "hi-base32": "^0.5.0",
+    "lodash": "^4.17.11",
+    "moment": "^2.22.2",
     "preact": "^8.2.6",
     "preact-cli-lodash": "^1.1.0",
     "preact-compat": "^3.17.0",
-    "react-modal": "^3.6.1",
     "preact-router": "^2.5.7",
     "qrcode.react": "^0.8.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "u2f-api": "^1.0.6",
+    "react-modal": "^3.6.1",
+    "ttag": "^1.4.2",
+    "u2f-api": "^1.0.9",
     "unistore": "^3.1.0"
   },
   "jest": {
